Provides classes that enable you to access ASP.NET forms authentication, roles, and profiles through the Figaro XML database library.
The Membership, Authentication, Role, and Profile classes contain
methods that let you log on users, check which roles the current user
belongs to, and retrieve profile properties for the user out of the
Figaro embedded XML database library.

This library is currently configured to use the Figaro DS x86 library,
though the source code will accommodate the CDS and TDS versions as well
(through build constants - look for the #if statements). All XQuery is
stored in resource files.